# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Registro Covid API",
        "description": "API do projeto Registro Covid",
        "contact": {
            "email": "email@test.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7000/api/",
            "description": "Registro Covid API"
        }
    ],
    "paths": {
...
        "/api/suportes-respiratorios": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista suportes respiratórios",
                "description": "Retorna todos os suportes respiratórios cadastrados no sistema",
                "operationId": "getSuporteRespiratorio",
                "responses": {
                    "200": {
                        "description": "Executado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "id": 1,
                                            "nome": "Máscara de reservatório"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
...
}